Changed basic file storage from nifti format to serialized JSON.
The nifti formated files require the nibabel module to be loaded upon page requests, which is less than ideal. In addition, since we are converting the image data into JSON on the fly, we are essentially wasting our time by doing so on each request. By pre-calculating the JSON and storing it in a raw format, we can serve the data much more quickly.