Added in caching mechanism to service.
After an image is rendered, it is stored in window.localStorage if it's available. However, this has introduced a bug. As we have to use a callback to execute the drawing of the cached image, there is a race condition between the brainRenderer and the crosshairsRenderer (which the crosshairs normally wins). This means that we see the crosshairs for a brief moment, before they are overdrawn by the loaded image from the cache.