Corrected crosshair rendering bug.
Refactored brainRenderers to take advantage of the rendering complete callback. The rendering of each stage then fires the completion callback, which will in turn fire the next renderer. This ensures that the rendering is done in the proper order (background, heatmap, then crosshairs) even though each of the renderers is operating in it's own thread that is spawned when it attempts to load a cached image. This also increases response time, but is giving a slight flicker on the redraw. We may want to reinvestigate this later.