Added in local page caching.
Images now attach themselves to the body in the DOM when they are first pulled from the cache. Thereafter when a cache hit occurs, that image is served from the DOM first, without the overhead of calling document.createElement everytime, then loading in the image data.

We had to alter getSlice in the brainRenderer to take a callback function to execute on image load, as is it the only function with knowledge of when an image will load. It's not a perfect solution, but the bridge is working well at this point.