Application now supports remote embeds.
A new script in the static dir is available, named brainvizEmbed.js. If you include this with a corresponding html/css template on a page, it will render the brain veiwer for you. We also removed some redunancies (onload.js) and moved some of that code into the embed script. Changed the code on the server side as well since the application now uses a special 'embed' script to embed itself in its own server. We'll need to find a good work around to get slice caching to work on remote systems, as img.toDataURL() does not work if that image comes from a remote server.